requirements.txt
fastapi~=0.116.1
uvicorn~=0.35.0
python-dotenv~=1.1.1
sqlalchemy~=2.0.41
psycopg2-binary==2.9.9
PyMuPDF
markdown
beautifulsoup4
openai~=1.97.1
python-multipart
faiss-cpu
tiktoken~=0.9.0
pydantic~=2.11.7
numpy~=2.0.2
pdfplumber